Revert "all features added"
