+++ src/App.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 
-
+import { AddressDetails } from './Components/Payment/Address'
 // import './App.css'
 import Products from './Components/Products/Products'
 import { Routes } from 'react-router-dom'
