+++ src/Components/AllRoutes.jsx
@@ -1,6 +1,6 @@
 import { Routes, Route } from "react-router-dom";
-import { PaymentDetails } from "./Payment/Payment";
-
+import { Paymentdetails } from "./Payment/Payment";
+import { AddressDetails } from "./Payment/Address";
 import Products from "./Products/Products";
 import ProductDetails from "./Products/ProductDetails";
 import HomePage from "./Home/HomePage";
@@ -23,8 +23,8 @@ export default function AllRoutes(){
 
         {/* <Route path="/login"/> */}
         {/* Poornima do edit below if required */}
-        <Route path="/payment" element={<PaymentDetails />}/>
-       
+        <Route path="/payment" element={<Paymentdetails />}/>
+        <Route path="/address" element={<AddressDetails />} />
         {/* Heena do edit if required below */}
         <Route path="/products" element={<Products />} />
         <Route path="/product/:id" element={<ProductDetails />} />
