+++ src/Components/Navbar/Navbar.module.css
@@ -357,6 +357,3 @@
     /* width: 50px; */
   }
 }
-
-
-
