--- src/Components/Payment/payment.module.css
@@ -1,130 +1,0 @@
-/* .shipping-form {
-    max-width: 400px;
-    margin: 20px auto;
-    padding: 20px;
-    border: 1px solid #ddd;
-    border-radius: 8px;
-    background-color: #fff;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  }
-  
-  .form-group {
-    margin-bottom: 20px;
-  }
-  
-  label {
-    display: block;
-    margin-bottom: 8px;
-    font-size: 14px;
-    color: #333;
-  }
-  
-  input {
-    width: 100%;
-    padding: 12px;
-    box-sizing: border-box;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    font-size: 14px;
-  }
-  
-  h2 {
-    font-size: 1.8em;
-    margin-bottom: 20px;
-    color: #333;
-  } */
-
-
-.addressbtn {
-
-  display: block;
-  width: 20%;
-  padding: 12px;
-  box-sizing: border-box;
-  border: none;
-  border-radius: 4px;
-  background-color: rgb(252, 205, 92);
-  color: black;
-  font-size: 16px;
-  cursor: pointer;
-  margin: auto;
-  font-weight: bold;
-
-}
-
-
-@media (max-width: 500px) {
-  .addressbtn {
-    max-width: 100%;
-    min-width: 150px;
-
-  }
-
-  .paymentbtn {
-    max-width: 100%;
-    min-width: 150px;
-    /* color: red; */
-  }
-
-  /*
-    .ppaygpayptm {
-      max-width: 100%;
-    }
-  
-    .debitCreditbtn {
-      max-width: 100%;
-    } */
-}
-
-
-.paymentoption{
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 20px;
-  justify-content: space-between;
-  height: 170px;
-  /* border: 2px solid red; */
-
-}
-.paytm{
-  width: 70px;
-  height: 70px;
-
-}
-.phonepay{
-height: 50px;
-border-radius: 5px;
-width: 50px;
-}
-@media (max-width:450px){
-  .paymentoption{
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 20px;
-    justify-content: space-between;
-    height: 150px;
-    /* border: 2px solid red; */
-  
-  }
-  .paytm{
-    height: 40px;
-    width: 40px;
-
-  }
-  .phonepay{
-    height: 30px;
-    width: 30px;
-  }
-}
-
-
-
-
-@media (max-width:570px) {
-  .paymentoption {
-    display: flex;
-    flex-wrap: wrap;
-  }
-}
