+++ src/main.jsx
@@ -9,13 +9,13 @@ import { BrowserRouter } from 'react-router-dom'
 import { Auth0Provider } from '@auth0/auth0-react';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
-   <Auth0Provider
+  <Auth0Provider
     domain="dev-tuhvv2ft848hmod6.us.auth0.com"
     clientId="w7IV3j2PXG0S4RDC3SxRqKpQN5NNAwyw"
     authorizationParams={{
       redirect_uri: window.location.origin
     }}
-   >
+  >
   <ChakraProvider>
   <Provider store={store}>
     <BrowserRouter>
@@ -23,5 +23,5 @@ ReactDOM.createRoot(document.getElementById('root')).render(
     </BrowserRouter>
     </Provider>
     </ChakraProvider>
-     </Auth0Provider>,
+    </Auth0Provider>,
 )
